refactor: 필요없는 부분 제거 및 책임의 맞는 위치이동(함수 이동,클래스 이동)
