+++ src/main/java/chess/Application.java
@@ -7,10 +7,6 @@ import chess.view.OutputView;
 public class Application {
     public static void main(String[] args) {
         ChessController controller = new ChessController(new InputView(), new OutputView());
-        try {
-            controller.run();
-        } catch (Exception exception) {
-            controller.printErrorMessage(exception);
-        }
+        controller.run();
     }
 }
